# backend/docs/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NutraCore API",
    "version": "1.1.0",
    "description": "Documentacion OpenAPI del backend NutraCore"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/api/reviews": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Crear o actualizar valoracion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Valoracion guardada"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/reviews/recipe/{recipeId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Listar valoraciones por receta",
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Valoraciones"
          }
        }
      }
    },
    "/api/reviews/{id}": {
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Eliminar valoracion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Valoracion eliminada"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ObjectId del recurso"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "No autorizado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "No encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "error": {
            "type": "string",
            "example": "Error de validacion"
          },
          "details": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          }
        }
      },
...
      "ReviewUpsertRequest": {
        "type": "object",
        "required": [
          "recipeId",
          "rating"
        ],
        "properties": {
          "recipeId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          }
        }
      },
...
    }
  }
}